+++ src/commands/gist.ts
@@ -17,6 +17,7 @@ import { duplicateGist, exportToRepo } from "../fileSystem/git";
 import { openRepo } from "../repos/store/actions";
 import { Gist, GistFile, GroupType, SortOrder, store } from "../store";
 import {
+  archiveGist,
   changeDescription,
   deleteGist,
   forkGist,
@@ -26,6 +27,7 @@ import {
   refreshGists,
   starGist,
   starredGists,
+  unarchiveGist,
   unstarGist
 } from "../store/actions";
 import { ensureAuthenticated, getApi, signIn } from "../store/auth";
@@ -36,7 +38,7 @@ import {
   GistsNode,
   StarredGistNode
 } from "../tree/nodes";
-import { createGistPadOpenUrl } from "../uriHandler";
+import { createGistPadOpenUrl, createGistPadWebUrl } from "../uriHandler";
 import {
   byteArrayToString,
   closeGistFiles,
@@ -46,6 +48,7 @@ import {
   getGistDetailsFromUri,
   getGistLabel,
   getGistWorkspaceId,
+  isArchivedGist,
   isGistWorkspace,
   isOwnedGist,
   openGist,
@@ -303,15 +306,26 @@ export async function registerGistCommands(context: ExtensionContext) {
         await ensureAuthenticated();
 
         if (node) {
-          const description = await window.showInputBox({
+          // If this is an archived gist, we need to remove the "archived" prefix
+          const isArchived = isArchivedGist(node.gist);
+
+          const description = isArchived ? node.gist.description.replace(" [Archived]", "") : node.gist.description;
+          let newDescription = await window.showInputBox({
             prompt: "Specify the description for this Gist",
-            value: node.gist.description
+            value: description
           });
 
-          if (!description) {
+          if (!newDescription) {
             return;
           }
-          await changeDescription(node.gist.id, description);
+
+          // If the gist was archived, we need to add the "archived" prefix
+          // back to the description, since we stripped it out above.
+          if (isArchived) {
+            newDescription += " [Archived]";
+          }
+
+          await changeDescription(node.gist.id, newDescription);
         }
       }
     )
@@ -343,7 +357,10 @@ export async function registerGistCommands(context: ExtensionContext) {
     commands.registerCommand(
       `${EXTENSION_NAME}.copyGistPadUrl`,
       async (node: GistNode) => {
-        const url = createGistPadOpenUrl(node.gist.id);
+        const url = node.gist.type === "note"
+          ? createGistPadWebUrl(node.gist.id)
+          : createGistPadOpenUrl(node.gist.id);
+
         env.clipboard.writeText(url);
       }
     )
@@ -486,11 +503,11 @@ export async function registerGistCommands(context: ExtensionContext) {
   );
 
   context.subscriptions.push(
-    commands.registerCommand(`${EXTENSION_NAME}.newPublicGist`, newPublicGist)
+    commands.registerCommand(`${EXTENSION_NAME}.newPublicGist`, () => newPublicGist())
   );
 
   context.subscriptions.push(
-    commands.registerCommand(`${EXTENSION_NAME}.newSecretGist`, newSecretGist)
+    commands.registerCommand(`${EXTENSION_NAME}.newSecretGist`, () => newSecretGist())
   );
 
   context.subscriptions.push(
@@ -523,7 +540,12 @@ export async function registerGistCommands(context: ExtensionContext) {
     commands.registerCommand(
       `${EXTENSION_NAME}.openGistInBrowser`,
       async (node: GistNode) => {
-        env.openExternal(Uri.parse(node.gist.html_url));
+        let url = node.gist.html_url;
+        if (node.gist.type === "note") {
+          url = createGistPadWebUrl(node.gist.id);
+        }
+
+        env.openExternal(Uri.parse(url));
       }
     )
   );
@@ -736,6 +758,26 @@ export async function registerGistCommands(context: ExtensionContext) {
             }
           }
         );
+      }
+    )
+  );
+
+  context.subscriptions.push(
+    commands.registerCommand(
+      `${EXTENSION_NAME}.archiveGist`,
+      async (node: GistNode) => {
+        await ensureAuthenticated();
+        await withProgress("Archiving gist...", () => archiveGist(node.gist.id));
+      }
+    )
+  );
+
+  context.subscriptions.push(
+    commands.registerCommand(
+      `${EXTENSION_NAME}.unarchiveGist`,
+      async (node: GistNode) => {
+        await ensureAuthenticated();
+        await withProgress("Unarchiving gist...", () => unarchiveGist(node.gist.id));
       }
     )
   );